requirements.txt
numpy==1.21.5
pandas==1.3.5
pytorch==1.12.1
streamlit==1.5.0
tqdm==4.62.3
ot==1.0.0